Move test fixtures out of extension root and fix CollectResumesButton test
Chrome MV3 rejects files/directories starting with "_". The __fixtures__
directory at the extension root caused:
"Cannot load extension with file or directory name __fixtures__"

Moved fixtures to src/lib/__tests__/__fixtures__/ and updated imports.

Also fixed CollectResumesButton test: window.open now uses named window
"trends-collect-seek" instead of "_blank".
