+++ apps/browser-extension/src/lib/__tests__/__fixtures__/job51-detail-mixed-with-projects.json
