+++ apps/browser-extension/src/lib/__tests__/__fixtures__/job51-detail-work-array.json
