+++ apps/browser-extension/src/lib/__tests__/job51-detail-parser.test.ts
@@ -1,8 +1,8 @@
 import { describe, expect, it, vi } from 'vitest'
 
-import mixedWithProjectsFixture from '../../../__fixtures__/job51-detail-mixed-with-projects.json'
-import workArrayFixture from '../../../__fixtures__/job51-detail-work-array.json'
-import workExperienceListFixture from '../../../__fixtures__/job51-detail-work-experience-list.json'
+import mixedWithProjectsFixture from './__fixtures__/job51-detail-mixed-with-projects.json'
+import workArrayFixture from './__fixtures__/job51-detail-work-array.json'
+import workExperienceListFixture from './__fixtures__/job51-detail-work-experience-list.json'
 import {
   EHIRE_51JOB_HOST,
   EHIRE_51JOB_PROFILE_URL_PREFIX,
